+++ tests/sim/test_material.py
@@ -7,10 +7,7 @@ from src.sim.material import gen_rand_material, Material, gen_rand_materials
 
 class TestMaterial(unittest.TestCase):
     def setUp(self):
-        pass
-
-    def tearDown(self):
-        pass
+        self.material = gen_rand_material()
 
     def test_eq(self):
         # arrange
@@ -79,7 +76,62 @@ class TestMaterial(unittest.TestCase):
         self.assertEqual(material, deepcopy_material)
         self.assertNotEqual(id(material), id(deepcopy_material))
 
-    def test_random_material(self):
+    def test_get_barcode(self):
+        # arrange
+        material = self.material
+
+        # act
+        barcode_get = material.get_barcode()
+        barcode_expected = material.barcode
+
+        # assert
+        self.assertEqual(barcode_get, barcode_expected)
+
+    def test_get_name(self):
+        # arrange
+        material = self.material
+
+        # act
+        name_get = material.get_name()
+        name_expected = material.name
+
+        # assert
+        self.assertEqual(name_get, name_expected)
+
+    def test_get_height(self):
+        # arrange
+        material = self.material
+
+        # act
+        height_get = material.get_height()
+        height_expected = material.height
+
+        # assert
+        self.assertEqual(height_get, height_expected)
+
+    def test_get_length(self):
+        # arrange
+        material = self.material
+
+        # act
+        length_get = material.get_length()
+        length_expected = material.length
+
+        # assert
+        self.assertEqual(length_get, length_expected)
+
+    def test_get_width(self):
+        # arrange
+        material = self.material
+
+        # act
+        width_get = material.get_width()
+        width_expected = material.width
+
+        # assert
+        self.assertEqual(width_get, width_expected)
+
+    def test_gen_rand_material_and_materials(self):
         # arrange
         max_height = 100
         minimum_height = 25